+++ src/main/java/de/gematik/demis/nrs/rules/model/RulesResultTypeEnum.java
@@ -38,16 +38,16 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public enum RulesResultTypeEnum {
   @JsonProperty("responsible_health_office")
-  RESPONSIBLE_HEALTH_OFFICE("responsible_health_office", 1),
+  RESPONSIBLE_HEALTH_OFFICE("responsible_health_office", 2),
   @JsonProperty("responsible_health_office_sormas")
   RESPONSIBLE_HEALTH_OFFICE_SORMAS("responsible_health_office_sormas", 4),
   @JsonProperty("responsible_health_office_with_relates_to")
-  RESPONSIBLE_HEALTH_OFFICE_WITH_RELATES_TO("responsible_health_office_with_relates_to", 3),
+  RESPONSIBLE_HEALTH_OFFICE_WITH_RELATES_TO("responsible_health_office_with_relates_to", 1),
   @JsonProperty("specific_receiver")
   SPECIFIC_RECEIVER("specific_receiver", 5),
   OTHER("", 99),
   @JsonProperty("responsible_health_office_tuberculosis")
-  RESPONSIBLE_HEALTH_OFFICE_TUBERCULOSIS("responsible_health_office_tuberculosis", 2);
+  RESPONSIBLE_HEALTH_OFFICE_TUBERCULOSIS("responsible_health_office_tuberculosis", 3);
 
   private static final Map<String, RulesResultTypeEnum> CODE_TO_ENUM =
       Arrays.stream(values())